requirements.txt
dlib
numpy
scikit-image
pandas
opencv-python
flask
pyttsx3
requests
pyinstaller
Pillow
